// packages/frontend/src/openapi/dnd-api/model/nonInteractiveEntityKind.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * DnD
 * The DnD API description
 * OpenAPI spec version: 1.0
 */

export type NonInteractiveEntityKind =
  (typeof NonInteractiveEntityKind)[keyof typeof NonInteractiveEntityKind];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NonInteractiveEntityKind = {
  wall: "wall",
  pillar: "pillar",
  tree: "tree",
  off_map: "off_map",
} as const;