// packages/frontend/src/openapi/dnd-api/model/tile.ts
import type { CoordResponseDto } from "./coordResponseDto";
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * DnD
 * The DnD API description
 * OpenAPI spec version: 1.0
 */
import type { TileEntitiesItem } from "./tileEntitiesItem";

export interface Tile {
  entities: TileEntitiesItem[];
  coord: CoordResponseDto;
  isStartingTile: boolean;
}