// packages/frontend/src/openapi/dnd-api/model/tileEntityType.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * DnD
 * The DnD API description
 * OpenAPI spec version: 1.0
 */

export type TileEntityType =
  (typeof TileEntityType)[keyof typeof TileEntityType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TileEntityType = {
  PLAYABLE_ENTITY: "PLAYABLE_ENTITY",
  INTERACTIVE_ENTITY: "INTERACTIVE_ENTITY",
  NON_INTERACTIVE_ENTITY: "NON_INTERACTIVE_ENTITY",
} as const;